feature: Adding Main Environmental Tasks and Up buttons
